This is my real JaveFX Testing Project aas the other repo is a bit f***** I hope this one works, and I hope I will be able to push my updates to this repo as this is a project to learn how to use JavaFX. It has simple examples.
I hope to be able to upgrade the login with the accountHandler and Account so that you can actually use it and log in.
And when I get that done, I am thinking about maybe upgrading it, so it uses JDBC for the accounts.

I have just made the JavaFX class "AccoutnScreen" Mostly to test the GitHub push thingy. I wanna see if this works...